+++ CHANGELOG.md
@@ -6,6 +6,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [1.2.3] - 2016-12-02
+### Changed
+- author could be string/array/object
+- react-native could be in dev-dependencoes
+### Added
+- inspect react
+
 ## [1.2.2] - 2016-11-24
 ### Added
 - inspect node & npm versions
@@ -29,8 +36,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - initialize this project
 
+[1.2.3]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.2...1.2.3
 [1.2.2]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.0...1.2.2
 [1.2.0]: https://github.com/super-fe/superfe-rn-inspector/compare/1.1.1...1.2.0
 [1.1.1]: https://github.com/super-fe/superfe-rn-inspector/compare/1.1.0...1.1.1
 [1.1.0]: https://github.com/super-fe/superfe-rn-inspector/compare/1.0.0...1.1.0
-[Unreleased]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.2...HEAD
+[Unreleased]: https://github.com/super-fe/superfe-rn-inspector/compare/1.2.3...HEAD